+++ addons/account_edi/models/__init__.py
@@ -2,9 +2,6 @@
 
 from . import account_move
 from . import account_journal
-from . import account_edi_format
-from . import account_edi_document
+from . import edi_format
+from . import edi_document
 from . import account_payment
-from . import ir_actions_report
-from . import mail_template
-from . import ir_attachment
