+++ addons/account_edi/models/account_journal.py
@@ -10,14 +10,14 @@ from collections import defaultdict
 class AccountJournal(models.Model):
     _inherit = 'account.journal'
 
-    edi_format_ids = fields.Many2many(comodel_name='account.edi.format',
+    edi_format_ids = fields.Many2many(comodel_name='edi.format',
                                       string='Electronic invoicing',
                                       help='Send XML/EDI invoices',
                                       domain="[('id', 'in', compatible_edi_ids)]",
                                       compute='_compute_edi_format_ids',
                                       readonly=False, store=True)
 
-    compatible_edi_ids = fields.Many2many(comodel_name='account.edi.format',
+    compatible_edi_ids = fields.Many2many(comodel_name='edi.format',
                                           compute='_compute_compatible_edi_ids',
                                           help='EDI format that support moves in this journal')
 
@@ -28,7 +28,7 @@ class AccountJournal(models.Model):
             old_edi_format_ids = self.edi_format_ids
             res = super().write(vals)
             diff_edi_format_ids = old_edi_format_ids - self.edi_format_ids
-            documents = self.env['account.edi.document'].search([
+            documents = self.env['edi.document'].search([
                 ('move_id.journal_id', 'in', self.ids),
                 ('edi_format_id', 'in', diff_edi_format_ids.ids),
                 ('state', 'in', ('to_cancel', 'to_send')),
@@ -41,7 +41,7 @@ class AccountJournal(models.Model):
 
     @api.depends('type', 'company_id', 'company_id.account_fiscal_country_id')
     def _compute_compatible_edi_ids(self):
-        edi_formats = self.env['account.edi.format'].search([])
+        edi_formats = self.env['edi.format'].search([])
 
         for journal in self:
             compatible_edis = edi_formats.filtered(lambda e: e._is_compatible_with_journal(journal))
@@ -49,7 +49,7 @@ class AccountJournal(models.Model):
 
     @api.depends('type', 'company_id', 'company_id.account_fiscal_country_id')
     def _compute_edi_format_ids(self):
-        edi_formats = self.env['account.edi.format'].search([])
+        edi_formats = self.env['edi.format'].search([])
         journal_ids = self.ids
 
         if journal_ids:
