+++ addons/account_edi/models/account_payment.py
@@ -5,7 +5,7 @@ from odoo import models, fields, api, _
 
 
 class AccountPayment(models.Model):
-    _inherit = 'account.payment'
+    _inherit = ['account.payment', 'edi']
 
     def action_process_edi_web_services(self):
         return self.move_id.action_process_edi_web_services()
