+++ addons/account_edi/security/ir.model.access.csv
@@ -1,5 +1,5 @@
 id,name,model_id:id,group_id:id,perm_read,perm_write,perm_create,perm_unlink
-access_account_edi_format_readonly,account.edi.format,model_account_edi_format,,1,0,0,0
-access_account_edi_format_group_invoice,account.edi.format,model_account_edi_format,account.group_account_invoice,1,1,1,1
-access_account_edi_document_readonly,account.edi.document,model_account_edi_document,,1,0,0,0
-access_account_edi_document_group_invoice,account.edi.document,model_account_edi_document,account.group_account_invoice,1,1,1,1
+access_account_edi_format_readonly,edi.format,model_account_edi_format,,1,0,0,0
+access_account_edi_format_group_invoice,edi.format,model_account_edi_format,account.group_account_invoice,1,1,1,1
+access_account_edi_document_readonly,edi.document,model_account_edi_document,,1,0,0,0
+access_account_edi_document_group_invoice,edi.document,model_account_edi_document,account.group_account_invoice,1,1,1,1
