+++ addons/account_edi_facturx/data/account_edi_data.xml
@@ -2,7 +2,7 @@
 <odoo>
     <data>
 
-        <record id="edi_facturx_1_0_05" model="account.edi.format">
+        <record id="edi_facturx_1_0_05" model="edi.format">
             <field name="name">Factur-X (FR)</field>
             <field name="code">facturx_1_0_05</field>
         </record>
