+++ addons/edi/__init__.py
@@ -1,0 +1,3 @@
+# -*- encoding: utf-8 -*-
+
+from . import models
