+++ addons/edi/models/__init__.py
@@ -1,0 +1,7 @@
+# -*- encoding: utf-8 -*-
+
+from . import edi_format
+from . import edi_document
+from . import ir_actions_report
+from . import mail_template
+from . import ir_attachment
