+++ addons/edi/models/edi_document.py
@@ -10,13 +10,13 @@ _logger = logging.getLogger(__name__)
 DEFAULT_BLOCKING_LEVEL = 'error'
 
 
-class AccountEdiDocument(models.Model):
-    _name = 'account.edi.document'
+class EdiDocument(models.Model):
+    _name = 'edi.document'
+    _inherits = {'ir.attachment': 'attachment_id'} # Idea is to support several kinds of relationships here
     _description = 'Electronic Document for an account.move'
 
     # == Stored fields ==
-    move_id = fields.Many2one('account.move', required=True, ondelete='cascade')
-    edi_format_id = fields.Many2one('account.edi.format', required=True)
+    edi_format_id = fields.Many2one('edi.format', required=True)
     attachment_id = fields.Many2one('ir.attachment', help='The file generated by edi_format_id when the invoice is posted (and this document is processed).')
     state = fields.Selection([('to_send', 'To Send'), ('sent', 'Sent'), ('to_cancel', 'To Cancel'), ('cancelled', 'Cancelled')])
     error = fields.Html(help='The text of the last error that happened during Electronic Invoice operation.')
@@ -31,13 +31,6 @@ class AccountEdiDocument(models.Model):
     name = fields.Char(related='attachment_id.name')
     edi_format_name = fields.Char(string='Format Name', related='edi_format_id.name')
 
-    _sql_constraints = [
-        (
-            'unique_edi_document_by_move_by_format',
-            'UNIQUE(edi_format_id, move_id)',
-            'Only one edi document by move by format',
-        ),
-    ]
 
     def _prepare_jobs(self):
         """Creates a list of jobs to be performed by '_process_job' for the documents in self.
@@ -51,38 +44,8 @@ class AccountEdiDocument(models.Model):
         """
 
         # Classify jobs by (edi_format, edi_doc.state, doc_type, move.company_id, custom_key)
-        to_process = {}
-        documents = self.filtered(lambda d: d.state in ('to_send', 'to_cancel') and d.blocking_level != 'error')
-        for edi_doc in documents:
-            move = edi_doc.move_id
-            edi_format = edi_doc.edi_format_id
-            if move.is_invoice(include_receipts=True):
-                doc_type = 'invoice'
-            elif move.payment_id or move.statement_line_id:
-                doc_type = 'payment'
-            else:
-                continue
-
-            custom_key = edi_format._get_batch_key(edi_doc.move_id, edi_doc.state)
-            key = (edi_format, edi_doc.state, doc_type, move.company_id, custom_key)
-            to_process.setdefault(key, self.env['account.edi.document'])
-            to_process[key] |= edi_doc
-
-        # Order payments/invoice and create batches.
-        invoices = []
-        payments = []
-        for key, documents in to_process.items():
-            edi_format, state, doc_type, company_id, custom_key = key
-            target = invoices if doc_type == 'invoice' else payments
-            batch = self.env['account.edi.document']
-            for doc in documents:
-                if edi_format._support_batching(move=doc.move_id, state=state, company=company_id):
-                    batch |= doc
-                else:
-                    target.append((doc, doc_type))
-            if batch:
-                target.append((batch, doc_type))
-        return invoices + payments
+        # TODO: Add res_model here?  (and then self.env[res_model].do_something() ?)
+        return {}
 
     @api.model
     def _process_job(self, documents, doc_type):
@@ -120,67 +83,7 @@ class AccountEdiDocument(models.Model):
             attachments_to_unlink.unlink()
 
         def _postprocess_cancel_edi_results(documents, edi_result):
-            invoice_ids_to_cancel = set()  # Avoid duplicates
-            attachments_to_unlink = self.env['ir.attachment']
-            for document in documents:
-                move = document.move_id
-                move_result = edi_result.get(move, {})
-                if move_result.get('success') is True:
-                    old_attachment = document.attachment_id
-                    document.write({
-                        'state': 'cancelled',
-                        'error': False,
-                        'attachment_id': False,
-                        'blocking_level': False,
-                    })
-
-                    if move.is_invoice(include_receipts=True) and move.state == 'posted':
-                        # The user requested a cancellation of the EDI and it has been approved. Then, the invoice
-                        # can be safely cancelled.
-                        invoice_ids_to_cancel.add(move.id)
-
-                    if not old_attachment.res_model or not old_attachment.res_id:
-                        attachments_to_unlink |= old_attachment
-
-                elif not move_result.get('success'):
-                    document.write({
-                        'error': move_result.get('error', False),
-                        'blocking_level': move_result.get('blocking_level', DEFAULT_BLOCKING_LEVEL) if move_result.get('error') else False,
-                    })
-
-            if invoice_ids_to_cancel:
-                invoices = self.env['account.move'].browse(list(invoice_ids_to_cancel))
-                invoices.button_draft()
-                invoices.button_cancel()
-
-            # Attachments that are not explicitly linked to a business model could be removed because they are not
-            # supposed to have any traceability from the user.
-            attachments_to_unlink.unlink()
-
-        test_mode = self._context.get('edi_test_mode', False)
-
-        documents.edi_format_id.ensure_one()  # All account.edi.document of a job should have the same edi_format_id
-        documents.move_id.company_id.ensure_one()  # All account.edi.document of a job should be from the same company
-        if len(set(doc.state for doc in documents)) != 1:
-            raise ValueError('All account.edi.document of a job should have the same state')
-
-        edi_format = documents.edi_format_id
-        state = documents[0].state
-        if doc_type == 'invoice':
-            if state == 'to_send':
-                edi_result = edi_format._post_invoice_edi(documents.move_id, test_mode=test_mode)
-                _postprocess_post_edi_results(documents, edi_result)
-            elif state == 'to_cancel':
-                edi_result = edi_format._cancel_invoice_edi(documents.move_id, test_mode=test_mode)
-                _postprocess_cancel_edi_results(documents, edi_result)
-
-        elif doc_type == 'payment':
-            if state == 'to_send':
-                edi_result = edi_format._post_payment_edi(documents.move_id, test_mode=test_mode)
-                _postprocess_post_edi_results(documents, edi_result)
-            elif state == 'to_cancel':
-                edi_result = edi_format._cancel_payment_edi(documents.move_id, test_mode=test_mode)
-                _postprocess_cancel_edi_results(documents, edi_result)
+            return {}
 
     def _process_documents_no_web_services(self):
         """ Post and cancel all the documents that don't need a web service.
