+++ addons/edi/models/mail_template.py
