+++ addons/edi/views/edi_document_views.xml
@@ -1,9 +1,9 @@
 <?xml version="1.0" encoding="utf-8"?>
 <odoo>
     <data>
-        <record id="view_tree_account_edi_document" model="ir.ui.view">
-            <field name="name">Account.edi.document.tree</field>
-            <field name="model">account.edi.document</field>
+        <record id="view_tree_edi_document" model="ir.ui.view">
+            <field name="name">edi.document.tree</field>
+            <field name="model">edi.document</field>
             <field name="arch" type="xml">
                 <tree create="false" delete="false" edit="false"
                       decoration-info="blocking_level == 'info'"
