+++ addons/l10n_be_edi/data/account_edi_data.xml
@@ -2,7 +2,7 @@
 <odoo>
     <data>
 
-        <record id="edi_efff_1" model="account.edi.format">
+        <record id="edi_efff_1" model="edi.format">
             <field name="name">E-FFF (BE)</field>
             <field name="code">efff_1</field>
         </record>
