+++ addons/l10n_be_edi/models/account_edi_format.py
@@ -6,7 +6,7 @@ import base64
 
 
 class AccountEdiFormat(models.Model):
-    _inherit = 'account.edi.format'
+    _inherit = 'edi.format'
 
     ####################################################
     # Export
@@ -31,7 +31,7 @@ class AccountEdiFormat(models.Model):
         })
 
     ####################################################
-    # Account.edi.format override
+    # edi.format override
     ####################################################
 
     def _create_invoice_from_xml_tree(self, filename, tree):
