+++ addons/l10n_it_edi/data/account_edi_data.xml
@@ -2,7 +2,7 @@
 <odoo>
     <data>
 
-        <record id="edi_fatturaPA" model="account.edi.format">
+        <record id="edi_fatturaPA" model="edi.format">
             <field name="name">Fattura PA (IT)</field>
             <field name="code">fattura_pa</field>
         </record>
