+++ addons/l10n_nl_edi/data/account_edi_data.xml
@@ -2,7 +2,7 @@
 <odoo>
     <data>
 
-        <record id="edi_nlcius_1" model="account.edi.format">
+        <record id="edi_nlcius_1" model="edi.format">
             <field name="name">NLCIUS (Netherlands)</field>
             <field name="code">nlcius_1</field>
         </record>
